refactor(ss-bench): reduce code redundance
