+++ rust/bench/ss_bench/operations/get.rs
@@ -1,6 +1,10 @@
 use std::time::Instant;
 
+use bytes::Bytes;
 use itertools::Itertools;
+use rand::distributions::Uniform;
+use rand::prelude::{Distribution, StdRng};
+use rand::SeedableRng;
 use risingwave_storage::StateStore;
 
 use super::Operations;
@@ -9,25 +13,13 @@ use crate::utils::workload::Workload;
 use crate::Opts;
 
 impl Operations {
-    pub(crate) async fn get_seq(&self, store: &impl StateStore, opts: &Opts) {
-        // generate queried point get key
-        let mut get_keys = match self.keys.is_empty() {
-            true => Workload::new_sequential_keys(opts).1,
-            false => {
-                assert!(
-                    opts.reads as usize <= self.keys.len(),
-                    "getseq cannot read more data than KV pairs in the state store"
-                );
-
-                (0..opts.reads as usize)
-                    .into_iter()
-                    .map(|i| self.keys[i].clone())
-                    .collect_vec()
-            }
-        };
-
+    async fn run_get(
+        store: &impl StateStore,
+        opts: &Opts,
+        mut get_keys: Vec<Bytes>,
+    ) -> (LatencyStat, u128, u128) {
         // partitioned these keys for each concurrency
-        let mut grouped_keys = vec![vec![]; opts.concurrency_num as usize];
+        let mut grouped_keys: Vec<Vec<Bytes>> = vec![vec![]; opts.concurrency_num as usize];
         for (i, key) in get_keys.drain(..).enumerate() {
             grouped_keys[i % opts.concurrency_num as usize].push(key);
         }
@@ -77,6 +69,53 @@ impl Operations {
         let stat = LatencyStat::new(total_latencies);
         let qps = opts.reads as u128 * 1_000_000_000 / total_time_nano as u128;
         let bytes_pre_sec = total_size as u128 * 1_000_000_000 / total_time_nano as u128;
+
+        (stat, qps, bytes_pre_sec)
+    }
+
+    pub(crate) async fn get_random(&self, store: &impl StateStore, opts: &Opts) {
+        // generate queried point get key
+        let get_keys = match self.keys.is_empty() {
+            true => Workload::new_random_keys(opts, 233).1,
+            false => {
+                let mut rng = StdRng::seed_from_u64(233);
+                let dist = Uniform::from(0..self.keys.len());
+                (0..opts.reads)
+                    .into_iter()
+                    .map(|_| self.keys[dist.sample(&mut rng)].clone())
+                    .collect_vec()
+            }
+        };
+
+        let (stat, qps, bytes_pre_sec) = Operations::run_get(store, opts, get_keys).await;
+
+        println!(
+            "
+    getrandom
+      {}
+      QPS: {}  {} bytes/sec",
+            stat, qps, bytes_pre_sec
+        );
+    }
+
+    pub(crate) async fn get_seq(&self, store: &impl StateStore, opts: &Opts) {
+        // generate queried point get key
+        let get_keys = match self.keys.is_empty() {
+            true => Workload::new_sequential_keys(opts).1,
+            false => {
+                assert!(
+                    opts.reads as usize <= self.keys.len(),
+                    "getseq cannot read more data than KV pairs in the state store"
+                );
+
+                (0..opts.reads as usize)
+                    .into_iter()
+                    .map(|i| self.keys[i].clone())
+                    .collect_vec()
+            }
+        };
+
+        let (stat, qps, bytes_pre_sec) = Operations::run_get(store, opts, get_keys).await;
 
         println!(
             "
